Split listener from sess.go
